+++ blog/src/main/java/com/jurgen/blog/dao/CommentDao.java
@@ -2,6 +2,6 @@ package com.jurgen.blog.dao;
 
 import com.jurgen.blog.domain.Comment;
 
-public interface CommentDao extends GenericDao<Comment, Integer> {
+public interface CommentDao extends GenericDao<Comment, Long> {
 
 }
