+++ blog/src/main/java/com/jurgen/blog/dao/RoleDao.java
@@ -2,7 +2,7 @@ package com.jurgen.blog.dao;
 
 import com.jurgen.blog.domain.UserRole;
 
-public interface RoleDao extends GenericDao<UserRole, Integer> {
+public interface RoleDao extends GenericDao<UserRole, Long> {
 
     public UserRole getRoleByName(String roleName);
 }
